DHFPROD-10237: I can't delete a values to ignore list after I removed it from my list.
