+++ marklogic-data-hub-central/ui/e2e/cypress/e2e/curation/curate/matching.cy.tsx
@@ -589,211 +589,7 @@ describe("Matching", () => {
     }
   });
 
-  it("Create new List in values to ignore", () => {
-    cy.visit("/tiles/curate");
-    cy.waitForAsyncRequest();
-    curatePage.toggleEntityTypeId("Person");
-    curatePage.selectMatchTab("Person");
-    curatePage.openStepDetails("match-person");
-    matchingStepDetail.addNewRuleset();
-    matchingStepDetail.getSinglePropertyOption();
 
-    cy.wait(1000);
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.createNewList();
-
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-    cy.findByText("A title for this list is required.");
-    cy.findByText("Values to ignore in this list are required.");
-
-
-    cy.log("**Create new List**");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "TitleList1");
-    rulesetSingleModal.addValuesToListToIgnore("Word1");
-    rulesetSingleModal.addValuesToListToIgnore("2Word");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-    cy.wait(500);
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    cy.findByText("TitleList1");
-
-    cy.log("**try to create with the same name of the last one list**");
-    rulesetSingleModal.createNewList();
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "TitleList1");
-    rulesetSingleModal.addValuesToListToIgnore("Word1");
-    rulesetSingleModal.addValuesToListToIgnore("2Word");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-
-    cy.findByText((_content, node) => {
-      const hasText = (node: Element) => node.textContent === "An existing list is already using the name TitleList1.";
-      const nodeHasText = hasText(node);
-      const childrenDontHaveText = Array.from(node.children).every(
-        (child) => !hasText(child)
-      );
-
-      return nodeHasText && childrenDontHaveText;
-    });
-
-    cy.log("**Create and close valid list**");
-    rulesetSingleModal.clearListTitle("values-to-ignore-input");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "TitleList1New");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-    cy.wait(1000);
-
-    cy.log("**Edit list of values to ignore**");
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.editListButton("TitleList1");
-    cy.findByText("Edit List");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "22");
-    rulesetSingleModal.addValuesToListToIgnore("Word3");
-    rulesetSingleModal.addValuesToListToIgnore("split word");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-    cy.wait(500);
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.hoverItemPresetList("TitleList122");
-    cy.findByText("Word1, 2Word, Word3, split word");
-
-    cy.log("**Copy List**");
-    rulesetSingleModal.copyListButton("TitleList122");
-    cy.findByText("List");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "MyCopy");
-    rulesetSingleModal.addValuesToListToIgnore("dog");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-    cy.wait(500);
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    cy.findByText("MyCopy");
-    rulesetSingleModal.hoverItemPresetList("MyCopy");
-    cy.findByText("Word1, 2Word, Word3, split word, dog");
-    rulesetSingleModal.closeButton().click();
-  });
-
-  it("Values to Ignore list tooltip", () => {
-    cy.visit("/tiles/curate");
-    cy.waitForAsyncRequest();
-
-    curatePage.toggleEntityTypeId("Person");
-    curatePage.selectMatchTab("Person");
-    curatePage.openStepDetails("match-person");
-    cy.findByLabelText("inputUriRadio").scrollIntoView().click();
-
-    cy.log("**adding new multiple property**");
-    cy.log("**check tooltip on presetList**");
-    matchingStepDetail.addNewRuleset();
-    matchingStepDetail.getSinglePropertyOption();
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    cy.wait(500);
-    if (Cypress.isBrowser("!chrome")) {
-      rulesetSingleModal.hoverItemPresetList("MyCopy");
-      cy.findByText("Word1, 2Word, Word3, split word, dog");
-      rulesetSingleModal.closeButton().click();
-    }
-  });
-
-  it("Select and save values to ignore", () => {
-    cy.visit("/tiles/curate");
-    cy.waitForAsyncRequest();
-
-    curatePage.toggleEntityTypeId("Person");
-    curatePage.selectMatchTab("Person");
-    curatePage.openStepDetails("match-person");
-    cy.findByLabelText("inputUriRadio").scrollIntoView().click();
-
-    cy.log("**Add To list to ignore**");
-    matchingStepDetail.addNewRuleset();
-    matchingStepDetail.getSinglePropertyOption();
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.createNewList();
-    rulesetSingleModal.addValuesToListToIgnore("Word1");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "swim");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.createNewList();
-    rulesetSingleModal.addValuesToListToIgnore("Word1");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "List1");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.selectItemFromList("swim");
-    rulesetSingleModal.selectItemFromList("List1");
-
-    rulesetSingleModal.selectPropertyToMatch("id");
-    rulesetSingleModal.selectMatchTypeDropdown("exact");
-    rulesetSingleModal.saveButton().click();
-    cy.waitForAsyncRequest();
-    cy.contains("id - Exact").should("have.length.gt", 0);
-    cy.findByLabelText("ruleset-scale-switch").click();
-    multiSlider.ruleSetEditOptionActive("id", "Exact");
-    cy.findByText("swim");
-    cy.findByText("List1");
-
-    cy.log("**Deleting a list with references**");
-    rulesetSingleModal.saveModalButton("cancel-single-ruleset");
-    matchingStepDetail.addNewRuleset();
-    matchingStepDetail.getSinglePropertyOption();
-    rulesetSingleModal.selectValuesToIgnoreInput();
-
-    rulesetSingleModal.deleteListButton("List1");
-    cy.waitForAsyncRequest();
-    cy.findAllByTestId("confirmation-modal").within(() => {
-      rulesetSingleModal.getElementByAriaLabel("Close").click();
-    });
-
-    cy.log("**Creating and deleting a new list without references**");
-    rulesetSingleModal.createNewList();
-    rulesetSingleModal.addValuesToListToIgnore("Word3");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "swim1");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.deleteListButton("swim1");
-    rulesetSingleModal.saveModalButton("Yes");
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.findText("swim1").should("not.exist");
-  });
-
-  it("Tests matching preview with values to ignore", () => {
-    cy.visit("/tiles/curate");
-    cy.waitForAsyncRequest();
-
-    curatePage.toggleEntityTypeId("Person");
-    curatePage.selectMatchTab("Person");
-    curatePage.openStepDetails("match-person");
-    matchingStepDetail.getAllDataRadio().click();
-    matchingStepDetail.getTestMatchUriButton();
-    cy.waitForAsyncRequest();
-    cy.wait(1000);
-    cy.findByLabelText("noMatchedDataView").should("not.exist");
-    let mergeResults = cy.get(`[id="testMatchedPanel"]`).first();
-    mergeResults.contains("Match - merge").should("have.length.gt", 0);
-    mergeResults.findByText("4 pair matches").should("have.length.gt", 0);
-    multiSlider.enableEdit("ruleset");
-    multiSlider.ruleSetEditOptionActive("lname", "Exact");
-    cy.wait(500);
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.createNewList();
-    rulesetSingleModal.addValuesToListToIgnore("Taylor");
-    rulesetSingleModal.addListTitle("values-to-ignore-input", "ignore-lastname");
-    rulesetSingleModal.saveModalButton("confirm-list-ignore");
-
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.selectItemFromList("ignore-lastname");
-    rulesetSingleModal.saveButton().click();
-    cy.waitForAsyncRequest();
-
-    matchingStepDetail.getAllDataRadio().click();
-    matchingStepDetail.getTestMatchUriButton();
-    cy.waitForAsyncRequest();
-    cy.wait(1000);
-    cy.findByLabelText("noMatchedDataView").should("not.exist");
-    mergeResults = cy.get(`[id="testMatchedPanel"]`).first();
-    mergeResults.contains("Match - merge").should("have.length.gt", 0);
-    mergeResults.findByText("3 pair matches").should("have.length.gt", 0);
-
-    multiSlider.ruleSetEditOptionActive("lname", "Exact");
-    cy.wait(500);
-    cy.get(`[id="valuesToIgnore"]`).type("{backspace}");
-    rulesetSingleModal.selectValuesToIgnoreInput();
-    rulesetSingleModal.saveButton().click();
-  });
 
   it("Check collection Typeahead request when source  database is changed", () => {
     cy.visit("/tiles/curate");
